Remove insertion of extra repairs in patch queue
Currently, if there is *any* repair in a patch queue, we insert a
repair after *every* patch that modifies the std schema.

This was in combination with deferring all repairs until the last one
in the queue, which allowed us to avoid needing to load the current
standard schema in order to do a repair. But we implemented support
for doing that as part of patching user extensions, so merge all of
the schema repair code into that.

I tested this on 5.x's substantial collection of complex patches, and
then forward ported. Since I had to resolve more conflicts there than
I expected, I then also injected a testing repair into the 6.x
backport (#8278), which shook out that `schema_repair()` is fully
broken on 6.x (fixed in #8279), but no patch-specific issues.