Add testing for the patch system
We test actually applying all of our real patches on release branches,
but this adds testing on the master branch of the patch system itself.

It works in basically the same way as our inplace upgrade tests: by
creating a database, and then applying a diff that makes some
synthetic testing changes to the server/compiler.

I (re)discovered some various frustrating limitations of the patch system
while doing this, but haven't fixed those yet.

Fixes #5804.