Also retry on transaction serialization errors
A lot of them popped up in the tests in the last two days, and we
should have been retrying them.

This is kind of weird, though because until quite recently we *never*
retried queries in `TRANSACTION_ISOLATION = False` tests and they
didn't fail *this* often.

The failures are all on things that interact with worker tasks
operating on the tables, though, so there is a lot of scope for timing
trouble.