Fix pg_catalog types over native SQL
Types such as pg_catalog.name or oid can be queried over native protocol,
but they don't have EdgeQL counterpart types.

This results in errors when we try to map pg oids to our types.
My solution is to add a few static mappings from known pg oids to our
type uuids.

This works for pg_catalog.name and oid, but does not for pg_catalog.aclitem.
If I try to spoof it as our `str`, it sends invalid UFT-8, which also causes
panic in the CLI (or rather rust bindings).