Rust implementation of JWT signing/validation (#8319)
Extracted from #8299

This implements JWT signing and validation in Rust.

The implementation is provided by the external crate `jsonwebtoken`, but
we manage all key loading and storage as that crate is somewhat lacking
in that department. The backend crypto for this is provided by Ring,
though the external crate may offer pluggable backends in the future.

The crate provides an interface `KeyRegistry` that can be used to
generate, load and save `JWK`s using one of three algorithms: `HS256`,
`RS256` or `EC256`.