Make \(name)-style string interpolation work for errmessage
This is (somewhat unfortunately) implemented by translating it to
the old {name} style used previously for errmessage.

It would be somewhat nicer to store the string using \(name)
style, but doing that loses the ability to prevent interpolation
by doing \\(name). (Since the lexed version is stored in that
case, we wouldn't be able to distinguish between the cases.)